--- src/main/kotlin/Lesson11_Task5/MessageForum.kt
@@ -1,8 +1,0 @@
-package Lesson11_Task5
-
-class MessageForum(val authorId: Int,
-                   val message: String,)
-
-{
-
-}
